fix(aegis): preserve startup and alert handoff
