+++ aegis/src/query.service.spec.ts
@@ -1,21 +1,9 @@
-import { ConfigService } from '@nestjs/config';
 import { Logger } from '@nestjs/common';
 import { register } from 'prom-client';
-import {
-  AbiEncodingLengthMismatchError,
-  ContractFunctionExecutionError,
-  ContractFunctionRevertedError,
-  createPublicClient,
-  http,
-  HttpRequestError,
-  RpcRequestError,
-} from 'viem';
-import { ChainConfig } from './config';
-import { Metric } from './metric';
+import { createPublicClient, http } from 'viem';
 import { QueryService } from './query.service';
+import { makeConfigService, makeMetric } from './query.service.test-fixtures';
 
-// Preserve viem's real error classes (HttpRequestError, etc.) so the source's
-// `instanceof` transport-error checks work; only stub the two functions we mock.
 jest.mock('viem', () => ({
   ...jest.requireActual('viem'),
   createPublicClient: jest.fn(),
@@ -27,114 +15,8 @@ jest.mock('viem', () => ({
 
 const mockCreatePublicClient = jest.mocked(createPublicClient);
 const mockHttp = jest.mocked(http);
-
-// A representative transport-level error (endpoint unreachable). Used where a
-// test needs the retry/counter path to fire.
-const makeTransportError = (message: string): HttpRequestError =>
-  new HttpRequestError({ url: 'http://localhost:8545', details: message });
-
-// A representative deterministic error (contract revert). viem wraps these as
-// ContractFunctionExecutionError → ContractFunctionRevertedError. Used where a
-// test needs to assert that no retry/counter path fires.
-const makeRevertError = (): ContractFunctionExecutionError => {
-  const cause = new ContractFunctionRevertedError({
-    abi: [],
-    functionName: 'getRateFeedTradingMode',
-    message: 'execution reverted',
-  });
-  return new ContractFunctionExecutionError(cause, {
-    abi: [
-      {
-        type: 'function',
-        name: 'getRateFeedTradingMode',
-        inputs: [],
-        outputs: [],
-        stateMutability: 'view',
-      },
-    ],
-    functionName: 'getRateFeedTradingMode',
-  });
-};
-
-// An RpcRequestError that carries an on-chain execution-revert code (3). Some
-// nodes surface this directly instead of viem's ContractFunctionRevertedError.
-const makeRpcRevertError = (): RpcRequestError =>
-  new RpcRequestError({
-    body: { method: 'eth_call', params: [] },
-    error: { code: 3, message: 'execution reverted' },
-    url: 'http://localhost:8545',
-  });
-
-// A transport failure wrapped inside ContractFunctionExecutionError, which is
-// exactly what viem's readContract emits when the *underlying cause* is a
-// network/HTTP error. The outer wrapper must NOT be treated as deterministic.
-const makeTransportWrappedInExecutionError =
-  (): ContractFunctionExecutionError =>
-    new ContractFunctionExecutionError(
-      new HttpRequestError({
-        url: 'http://localhost:8545',
-        details: 'connection refused',
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      }) as any,
-      {
-        abi: [
-          {
-            type: 'function',
-            name: 'getRateFeedTradingMode',
-            inputs: [],
-            outputs: [],
-            stateMutability: 'view',
-          },
-        ],
-        functionName: 'getRateFeedTradingMode',
-      },
-    );
-
 const readContract = jest.fn();
 const getBalance = jest.fn();
-
-const chain = {
-  id: 'localnet',
-  label: 'Localnet',
-  httpRpcUrl: 'http://localhost:8545',
-  contracts: {
-    BreakerBox: '0x0000000000000000000000000000000000000001',
-  },
-  vars: {
-    FeedId: '0xfeed',
-  },
-} as unknown as ChainConfig;
-
-const makeConfigService = (
-  chains: ChainConfig[] | undefined = [chain],
-  hasChains = true,
-): jest.Mocked<ConfigService> =>
-  ({
-    get: jest.fn((key: string) => {
-      if (key === 'chains') return hasChains ? chains : undefined;
-      return undefined;
-    }),
-  }) as unknown as jest.Mocked<ConfigService>;
-
-const makeMetric = (overrides: Record<string, unknown> = {}): Metric =>
-  ({
-    chain: 'localnet',
-    name: 'BreakerBox_getRateFeedTradingMode',
-    source: {
-      contract: 'BreakerBox',
-      raw: 'BreakerBox.getRateFeedTradingMode(bytes32 rateFeedId)(uint8 mode)',
-      functionAbi: {
-        type: 'function',
-        name: 'getRateFeedTradingMode',
-        stateMutability: 'view',
-        inputs: [{ type: 'bytes32', name: 'rateFeedId' }],
-        outputs: [{ type: 'uint8', name: 'mode' }],
-      },
-    },
-    args: ['FeedId', 'literal'],
-    parse: jest.fn(() => 1),
-    ...overrides,
-  }) as unknown as Metric;
 
 describe('QueryService', () => {
   beforeEach(() => {
@@ -224,8 +106,7 @@ describe('QueryService', () => {
   });
 
   it('returns undefined and records an error path when a view call fails', async () => {
-    const error = new Error('rpc unavailable');
-    readContract.mockRejectedValue(error);
+    readContract.mockRejectedValue(new Error('rpc unavailable'));
     const loggerError = jest
       .spyOn(Logger.prototype, 'error')
       .mockImplementation(() => undefined);
@@ -270,398 +151,5 @@ describe('QueryService', () => {
       'Missing function name for metric BreakerBox_getRateFeedTradingMode',
     );
     expect(readContract).not.toHaveBeenCalled();
-  });
-
-  // (A) Counter increments on a transport failure with no fallback
-  it('increments rpcErrors counter when a transport call fails and no fallback is configured', async () => {
-    readContract.mockRejectedValue(makeTransportError('rpc unavailable'));
-    jest.spyOn(Logger.prototype, 'error').mockImplementation(() => undefined);
-    const service = new QueryService(makeConfigService());
-    const metric = makeMetric();
-
-    await expect(service.query(metric)).resolves.toBeUndefined();
-
-    const metrics = await service.rpcErrors.get();
-    expect(metrics.values).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          labels: {
-            contract: 'BreakerBox',
-            functionName: 'getRateFeedTradingMode',
-            chain: 'localnet',
-          },
-          value: 1,
-        }),
-      ]),
-    );
-  });
-
-  // (B) Fallback used when primary fails with a transport error
-  it('retries via fallback client when primary RPC fails with a transport error', async () => {
-    const primaryReadContract = jest
-      .fn()
-      .mockRejectedValue(makeTransportError('primary down'));
-    const fallbackReadContract = jest.fn().mockResolvedValue(42n);
-
-    mockCreatePublicClient
-      .mockReturnValueOnce({
-        readContract: primaryReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>)
-      .mockReturnValueOnce({
-        readContract: fallbackReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>);
-
-    jest.spyOn(Logger.prototype, 'warn').mockImplementation(() => undefined);
-
-    const chainWithFallback = {
-      ...chain,
-      fallbackHttpRpcUrl: 'http://localhost:8546',
-    } as unknown as ChainConfig;
-    const service = new QueryService(makeConfigService([chainWithFallback]));
-    const metric = makeMetric({ parse: jest.fn(() => 42) });
-
-    await expect(service.query(metric)).resolves.toBe(42);
-
-    expect(primaryReadContract).toHaveBeenCalledTimes(1);
-    expect(fallbackReadContract).toHaveBeenCalledTimes(1);
-  });
-
-  // (C) Counter NOT incremented when fallback succeeds
-  it('does not increment rpcErrors counter when fallback succeeds', async () => {
-    const primaryReadContract = jest
-      .fn()
-      .mockRejectedValue(makeTransportError('primary down'));
-    const fallbackReadContract = jest.fn().mockResolvedValue(42n);
-
-    mockCreatePublicClient
-      .mockReturnValueOnce({
-        readContract: primaryReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>)
-      .mockReturnValueOnce({
-        readContract: fallbackReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>);
-
-    jest.spyOn(Logger.prototype, 'warn').mockImplementation(() => undefined);
-
-    const chainWithFallback = {
-      ...chain,
-      fallbackHttpRpcUrl: 'http://localhost:8546',
-    } as unknown as ChainConfig;
-    const service = new QueryService(makeConfigService([chainWithFallback]));
-    const metric = makeMetric({ parse: jest.fn(() => 42) });
-
-    await expect(service.query(metric)).resolves.toBe(42);
-
-    const metrics = await service.rpcErrors.get();
-    const total = metrics.values.reduce((sum, v) => sum + v.value, 0);
-    expect(total).toBe(0);
-  });
-
-  // (D) Counter incremented when both primary and fallback fail (transport)
-  it('increments rpcErrors counter when both primary and fallback fail', async () => {
-    const primaryError = makeTransportError('primary down');
-    const fallbackError = makeTransportError('fallback down');
-    const primaryReadContract = jest.fn().mockRejectedValue(primaryError);
-    const fallbackReadContract = jest.fn().mockRejectedValue(fallbackError);
-
-    mockCreatePublicClient
-      .mockReturnValueOnce({
-        readContract: primaryReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>)
-      .mockReturnValueOnce({
-        readContract: fallbackReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>);
-
-    const loggerWarn = jest
-      .spyOn(Logger.prototype, 'warn')
-      .mockImplementation(() => undefined);
-    const loggerError = jest
-      .spyOn(Logger.prototype, 'error')
-      .mockImplementation(() => undefined);
-
-    const chainWithFallback = {
-      ...chain,
-      fallbackHttpRpcUrl: 'http://localhost:8546',
-    } as unknown as ChainConfig;
-    const service = new QueryService(makeConfigService([chainWithFallback]));
-    const metric = makeMetric();
-
-    await expect(service.query(metric)).resolves.toBeUndefined();
-
-    // The warn fires before the fallback is attempted and carries the primary error.
-    expect(loggerWarn).toHaveBeenCalledWith(
-      expect.stringContaining(
-        'Primary RPC failed for BreakerBox.getRateFeedTradingMode on localnet',
-      ),
-    );
-    // The concise viem message omits request and response payloads.
-    expect(loggerWarn).toHaveBeenCalledWith(
-      expect.stringContaining('HTTP request failed.'),
-    );
-    // When both fail, the terminal log identifies both endpoint outcomes.
-    expect(loggerError).toHaveBeenCalledWith(
-      expect.stringContaining('Primary: HTTP request failed.'),
-    );
-    expect(loggerError).toHaveBeenCalledWith(
-      expect.stringContaining('Fallback: HTTP request failed.'),
-    );
-
-    const metrics = await service.rpcErrors.get();
-    expect(metrics.values).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          labels: {
-            contract: 'BreakerBox',
-            functionName: 'getRateFeedTradingMode',
-            chain: 'localnet',
-          },
-          value: 1,
-        }),
-      ]),
-    );
-  });
-
-  it('logs primary-only and dual-endpoint terminal failures independently', async () => {
-    const primaryReadContract = jest
-      .fn()
-      .mockRejectedValueOnce(makeRevertError())
-      .mockRejectedValueOnce(makeTransportError('primary down'));
-    const fallbackReadContract = jest
-      .fn()
-      .mockRejectedValue(makeTransportError('fallback down'));
-
-    mockCreatePublicClient
-      .mockReturnValueOnce({
-        readContract: primaryReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>)
-      .mockReturnValueOnce({
-        readContract: fallbackReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>);
-
-    jest.spyOn(Logger.prototype, 'warn').mockImplementation(() => undefined);
-    const loggerError = jest
-      .spyOn(Logger.prototype, 'error')
-      .mockImplementation(() => undefined);
-    const chainWithFallback = {
-      ...chain,
-      fallbackHttpRpcUrl: 'http://localhost:8546',
-    } as unknown as ChainConfig;
-    const service = new QueryService(makeConfigService([chainWithFallback]));
-    const metric = makeMetric();
-
-    await expect(service.query(metric)).resolves.toBeUndefined();
-    await expect(service.query(metric)).resolves.toBeUndefined();
-
-    expect(loggerError).toHaveBeenCalledTimes(2);
-    expect(loggerError).toHaveBeenNthCalledWith(
-      1,
-      expect.stringContaining('RPC call failed for'),
-    );
-    expect(loggerError).toHaveBeenNthCalledWith(
-      2,
-      expect.stringContaining('Both RPC endpoints failed for'),
-    );
-  });
-
-  // (E) Counter NOT incremented when the RPC succeeds but parse throws.
-  // The counter tracks RPC-transport failures only, not parse/validation errors.
-  it('does not increment rpcErrors counter when parse fails after a successful RPC', async () => {
-    readContract.mockResolvedValue(2n);
-    jest.spyOn(Logger.prototype, 'error').mockImplementation(() => undefined);
-    const service = new QueryService(makeConfigService());
-    const metric = makeMetric({
-      parse: jest.fn(() => {
-        throw new Error('unparseable value');
-      }),
-    });
-
-    await expect(service.query(metric)).resolves.toBeUndefined();
-
-    const metrics = await service.rpcErrors.get();
-    const total = metrics.values.reduce((sum, v) => sum + v.value, 0);
-    expect(total).toBe(0);
-  });
-
-  // (F) A deterministic call failure (contract revert — ContractFunctionRevertedError
-  // nested inside ContractFunctionExecutionError) is NOT a transport error: the
-  // fallback must not be tried and the counter must stay flat.
-  it('does not retry fallback or increment counter on a contract revert (ContractFunctionExecutionError)', async () => {
-    const primaryReadContract = jest.fn().mockRejectedValue(makeRevertError());
-    const fallbackReadContract = jest.fn().mockResolvedValue(99n);
-
-    mockCreatePublicClient
-      .mockReturnValueOnce({
-        readContract: primaryReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>)
-      .mockReturnValueOnce({
-        readContract: fallbackReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>);
-
-    jest.spyOn(Logger.prototype, 'warn').mockImplementation(() => undefined);
-    jest.spyOn(Logger.prototype, 'error').mockImplementation(() => undefined);
-
-    const chainWithFallback = {
-      ...chain,
-      fallbackHttpRpcUrl: 'http://localhost:8546',
-    } as unknown as ChainConfig;
-    const service = new QueryService(makeConfigService([chainWithFallback]));
-    const metric = makeMetric();
-
-    await expect(service.query(metric)).resolves.toBeUndefined();
-
-    // Fallback must NOT be attempted for a deterministic error.
-    expect(fallbackReadContract).not.toHaveBeenCalled();
-
-    // Counter must stay flat — this is a protocol/config failure, not an outage.
-    const metrics = await service.rpcErrors.get();
-    const total = metrics.values.reduce((sum, v) => sum + v.value, 0);
-    expect(total).toBe(0);
-  });
-
-  // (G-H order): G = RpcRequestError revert, H = transport wrapped in exec error.
-
-  // (G) An RpcRequestError with code 3 (execution reverted) is deterministic:
-  // no fallback retry, no counter increment.
-  it('does not retry fallback or increment counter on an RpcRequestError with revert code', async () => {
-    const primaryReadContract = jest
-      .fn()
-      .mockRejectedValue(makeRpcRevertError());
-    const fallbackReadContract = jest.fn().mockResolvedValue(99n);
-
-    mockCreatePublicClient
-      .mockReturnValueOnce({
-        readContract: primaryReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>)
-      .mockReturnValueOnce({
-        readContract: fallbackReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>);
-
-    jest.spyOn(Logger.prototype, 'warn').mockImplementation(() => undefined);
-    jest.spyOn(Logger.prototype, 'error').mockImplementation(() => undefined);
-
-    const chainWithFallback = {
-      ...chain,
-      fallbackHttpRpcUrl: 'http://localhost:8546',
-    } as unknown as ChainConfig;
-    const service = new QueryService(makeConfigService([chainWithFallback]));
-    const metric = makeMetric();
-
-    await expect(service.query(metric)).resolves.toBeUndefined();
-
-    expect(fallbackReadContract).not.toHaveBeenCalled();
-
-    const metrics = await service.rpcErrors.get();
-    const total = metrics.values.reduce((sum, v) => sum + v.value, 0);
-    expect(total).toBe(0);
-  });
-
-  // (H) A transport failure (HttpRequestError) wrapped inside viem's
-  // ContractFunctionExecutionError — the exact shape readContract emits for
-  // network-level errors — IS a transport error: the fallback must be tried and
-  // the counter must increment when all endpoints fail.
-  it('retries fallback and increments counter when a transport error is wrapped in ContractFunctionExecutionError', async () => {
-    const transportWrapped = makeTransportWrappedInExecutionError();
-    const primaryReadContract = jest.fn().mockRejectedValue(transportWrapped);
-    const fallbackReadContract = jest
-      .fn()
-      .mockRejectedValue(makeTransportError('fallback down'));
-
-    mockCreatePublicClient
-      .mockReturnValueOnce({
-        readContract: primaryReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>)
-      .mockReturnValueOnce({
-        readContract: fallbackReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>);
-
-    jest.spyOn(Logger.prototype, 'warn').mockImplementation(() => undefined);
-    jest.spyOn(Logger.prototype, 'error').mockImplementation(() => undefined);
-
-    const chainWithFallback = {
-      ...chain,
-      fallbackHttpRpcUrl: 'http://localhost:8546',
-    } as unknown as ChainConfig;
-    const service = new QueryService(makeConfigService([chainWithFallback]));
-    const metric = makeMetric();
-
-    await expect(service.query(metric)).resolves.toBeUndefined();
-
-    // The fallback MUST be attempted — ContractFunctionExecutionError wrapping a
-    // transport cause is NOT deterministic.
-    expect(fallbackReadContract).toHaveBeenCalledTimes(1);
-
-    // Both failed with transport errors, so the counter must increment.
-    const metrics = await service.rpcErrors.get();
-    expect(metrics.values).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          labels: {
-            contract: 'BreakerBox',
-            functionName: 'getRateFeedTradingMode',
-            chain: 'localnet',
-          },
-          value: 1,
-        }),
-      ]),
-    );
-  });
-
-  // (I) An ABI encoding error (e.g. wrong number of args for a function
-  // signature) is deterministic — it would reproduce on every healthy endpoint.
-  // The fallback must NOT be retried and the counter must stay flat.
-  it('does not retry fallback or increment counter on an ABI encoding error', async () => {
-    // AbiEncodingLengthMismatchError is what viem raises when the number of
-    // encoded values does not match the ABI param list — a typical outcome of a
-    // bad metric ABI / arg-shape configuration.
-    const abiError = new AbiEncodingLengthMismatchError({
-      expectedLength: 1,
-      givenLength: 0,
-    });
-    const primaryReadContract = jest.fn().mockRejectedValue(abiError);
-    const fallbackReadContract = jest.fn().mockResolvedValue(99n);
-
-    mockCreatePublicClient
-      .mockReturnValueOnce({
-        readContract: primaryReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>)
-      .mockReturnValueOnce({
-        readContract: fallbackReadContract,
-        getBalance: jest.fn(),
-      } as unknown as ReturnType<typeof createPublicClient>);
-
-    jest.spyOn(Logger.prototype, 'warn').mockImplementation(() => undefined);
-    jest.spyOn(Logger.prototype, 'error').mockImplementation(() => undefined);
-
-    const chainWithFallback = {
-      ...chain,
-      fallbackHttpRpcUrl: 'http://localhost:8546',
-    } as unknown as ChainConfig;
-    const service = new QueryService(makeConfigService([chainWithFallback]));
-    const metric = makeMetric();
-
-    await expect(service.query(metric)).resolves.toBeUndefined();
-
-    // Fallback must NOT be attempted — the ABI error is deterministic.
-    expect(fallbackReadContract).not.toHaveBeenCalled();
-
-    // Counter must stay flat — this is a config/ABI failure, not an outage.
-    const metrics = await service.rpcErrors.get();
-    const total = metrics.values.reduce((sum, v) => sum + v.value, 0);
-    expect(total).toBe(0);
   });
 });
